pyproject.toml
[project]
name = "nba-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "ipykernel>=7.1.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.0",
    "uvicorn>=0.34.0",
    "xgboost>=2.0.0",
]
